[F!] Callback menu defs are now deferred
This has the potential to be backwards in-compatible, because
now your functions may be called at a different time than you expect.
If that's the case, you were probably working around the fact that
these functions weren't called late enough (like after other code
has been initialized). So, this fixes that!